e2e/requirements.txt
albumentations
cvxpy
hdbscan
imageio[ffmpeg]
ipython
lightning
matplotlib
numba
numpy
opencv-python
pandas
pytest
pytorch-msssim
PyWavelets
scikit-image
scikit-learn
scipy
seaborn
torchvision
tqdm
